requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
django-filter==2.4.0
djangorestframework-simplejwt==4.5.0
psycopg2-binary==2.8.6
